% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{downloadUKterrain}
\alias{downloadUKterrain}
\title{Download data from the Shuttle Radar Topography Mission for the UK}
\usage{
downloadUKterrain(mainDir)
}
\arguments{
\item{Defaults}{to TRUE.}
}
\description{
This function allows you to conveniently download UK terrain data from the
SRTM3 dataset. This is the highest resolution SRTM dataset currenltly
available and is particularly useful for running air qulity models.
}
\details{
To run the downloader run the function with the desired download location
as the only argument. This will place a copy of all UK SRTM3 files in that
directory.
}
\examples{
downloadUKterrain("/Users/Scott/Downloads/")

}
\keyword{terrain}

