% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mapterrain}
\alias{mapterrain}
\title{Prepare a simple map with a terrain tile from the SRTM3 dataset using
leaflet. To run the function simply pass the desired .hgt file to the
function as an argument.}
\usage{
mapterrain(tile)
}
\arguments{
\item{Defaults}{to TRUE.}
}
\description{
Prepare a simple map with a terrain tile from the SRTM3 dataset using
leaflet. To run the function simply pass the desired .hgt file to the
function as an argument.
}
\examples{
mapterrain("/Users/Scott/Downloads/N55W005.hgt")

}
\keyword{terrain}

