% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{showUKterrain}
\alias{showUKterrain}
\title{Show the coverage of the SRTM3 datasets for the UK in a leaflet map.}
\usage{
showUKterrain()
}
\arguments{
\item{Defaults}{to TRUE.}
}
\description{
This convenience functions shows the locations of the SRTM3 data tiles in
the UK. Each point in the map denotes the origin of the tile at the south
west corner. The function has no arguments, so run with empty brackets.
}
\examples{
showUKterrain()

}
\keyword{terrain}

