% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{toUKterrain}
\alias{toUKterrain}
\title{Convert SRTM3 dataset to the OS grid coordinate reference system EPSG 27700.}
\usage{
toUKterrain(filepath, tile)
}
\arguments{
\item{Defaults}{to TRUE.}
}
\description{
To run the converter run the function with the desired file for projection
to the OS grid coordinate system as the only argument.
}
\examples{
toUKterrain("/Users/Scott/Downloads/N55W005.hgt")

}
\keyword{terrain}

